auto {flexclust}	R Documentation

Automobile Customer Survey Data

Description

A German manufacturer of premium cars asked customers approximately 3 months after a car purchase which characteristics of the car were most important for the decision to buy the car. The survey was done in 1983 and the data set contains all responses without missing values.

Usage

data(auto)

Format

A data frame with 793 observations on the following 46 variables.

model

a factor with levels A B C D, model bought by the customer.

gear

a factor with levels 4 gears, 5 econo, 5 sport, or automatic.

leasing

a logical vector, was leasing used to finance the car?

usage

a factor with levels private, both, business.

previous_model

a factor describing which type of car was owned directly before the purchase.

other_consider

a factor with levels same manuf, other manuf, both, or none

test_drive

a logical vector, did you do a test drive?

info_adv

a logical vector, was advertising an important source of information?

info_exp

a logical vector, was experience an important source of information?

info_rec

a logical vector, were recommendations an important source of information?

ch_clarity

a logical vector

ch_economy

a logical vector

ch_driving_properties

a logical vector

ch_service

a logical vector

ch_interior

a logical vector

ch_quality

a logical vector

ch_technology

a logical vector

ch_model_continuity

a logical vector

ch_comfort

a logical vector

ch_reliability

a logical vector

ch_handling

a logical vector

ch_reputation

a logical vector

ch_concept

a logical vector

ch_character

a logical vector

ch_power

a logical vector

ch_resale_value

a logical vector

ch_styling

a logical vector

ch_safety

a logical vector

ch_sporty

a logical vector

ch_consumption

a logical vector

ch_space

a logical vector

satisfaction

a numeric vector describing overall satisfaction (1=very good, 10=very bad).

good1

conception, styling, dimensions.

good2

auto body.

good3

driving and coupled axles.

good4

engine.

good5

electronics.

good6

financing and customer service.

good7

other.

sporty

What do you think about the balance of sportiness and comfort? (good more sport more comfort).

drive_char

driving characteristis (gentle < speedy < powerfull < extreme).

tempo

Which average speed do you prefer on German Autobahn in km/h? (< 130 < 130-150 < 150-180 < > 180)

consumption

an ordered factor with levels low < ok < high < too high.

gender

a factor with levels male female

occupation

a factor with levels self-employed, freelance, and employee.

household

size of household, an ordered factor with levels 1-2 < >=3

Source

The original German data are in the public domain and available at http://dx.doi.org/10.5282/ubm/data.14 from LMU Munich. The variable names and help page were translated to English and converted into Rd format by Friedrich Leisch.

References

Open Data LMU (1983): Umfrage unter Kunden einer Automobilfirma, doi:10.5282/ubm/data.14

Examples

data(auto)
summary(auto)

[Package flexclust version 1.3-5 Index]