flxColors {flexclust}	R Documentation

Flexclust Color Palettes

Description

Create and access palettes for the plot methods.

Usage

  flxColors(n=1:8, color=c("full","medium", "light","dark"), grey=FALSE)

Arguments

n
Index number of color to return (1 to 8)

color
Type of color, see details.

grey
Return grey value corresponding to palette.

Details

This function creates color palettes in HCL space for up to 8 colors. All palettes have constant chroma and luminance, only the hue of the colors change within a palette.

Palettes "full" and "dark" have the same luminance, and palettes "medium" and "light" have the same luminance.

Author(s)

Friedrich Leisch

See Also

hcl

Examples

opar <- par(c("mfrow", "mar", "xaxt"))
par(mfrow=c(2,2), mar=c(0,0,2,0), yaxt="n")

x <- rep(1, 8)

barplot(x, col = flxColors(color="full"), main="full")
barplot(x, col = flxColors(color="dark"), main="dark")
barplot(x, col = flxColors(color="medium"), main="medium")
barplot(x, col = flxColors(color="light"), main="light")

par(opar)

[Package flexclust version 1.3-5 Index]