image-methods {flexclust}	R Documentation

Methods for Function image in Package ‘flexclust’

Description

Image plot of cluster segments overlaid by neighbourhood graph.

Usage

## S4 method for signature 'kcca'
image(x, which = 1:2, npoints = 100,
         xlab = "", ylab = "", fastcol = TRUE, col=NULL,
         clwd=0, graph=TRUE, ...)

Arguments

x
An object of class "kcca".

which
Index number of dimensions of input space to plot.

npoints
Number of grid points for image.

fastcol
If TRUE, a greedy algorithm is used for the background colors of the segments, which may result in neighbouring segments having the same color. If FALSE, neighbouring segments always have different colors at a speed penalty.

col
Vector of background colors for the segments.

clwd
Line width of contour lines at cluster boundaries, use larger values for npoints than the default to get smooth lines. (Warning: We really need a smarter way to draw cluster boundaries!)

graph
Logical, add a neighborhood graph to the plot?

xlab, ylab, ...
Graphical parameters.

Details

This works only for "kcca" objects, no method is available for "kccasimple" objects.

Author(s)

Friedrich Leisch

See Also

kcca

[Package flexclust version 1.3-5 Index]