info {flexclust}	R Documentation

Get Information on Fitted Flexclust Objects

Description

Returns descriptive information about fitted flexclust objects like cluster sizes or sum of within-cluster distances.

Usage

## S4 method for signature 'flexclust,character'
info(object, which, drop=TRUE, ...)

Arguments

object
fitted object.

which
which information to get. Use which="help" to list available information.

drop
logical. If TRUE the result is coerced to the lowest possible dimension.

...
passed to methods.

Details

Function info can be used to access slots of fitted flexclust objects in a portable way, and in addition computes some meta-information like sum of within-cluster distances.

Function infoCheck returns a logical value that is TRUE if the requested information can be computed from the object.

Author(s)

Friedrich Leisch

See Also

info

Examples

data("Nclus")
plot(Nclus)

cl1 = cclust(Nclus, k=4)
summary(cl1)

## these two are the same
info(cl1)
info(cl1, "help")

## cluster sizes
i1 = info(cl1, "size")
i1

## average within cluster distances
i2 = info(cl1, "av_dist")
i2

## the sum of all within-cluster distances
i3 = info(cl1, "distsum")
i3

## sum(i1*i2) must of course be the same as i3
stopifnot(all.equal(sum(i1*i2), i3))

## This should return TRUE
infoCheck(cl1, "size")
## and this FALSE
infoCheck(cl1, "Homer Simpson")
## both combined
i4 = infoCheck(cl1, c("size", "Homer Simpson"))
i4

stopifnot(all.equal(i4, c(TRUE, FALSE)))

[Package flexclust version 1.3-5 Index]