pairs-methods {flexclust}	R Documentation

Methods for Function pairs in Package ‘flexclust’

Description

Plot a matrix of neighbourhood graphs.

Usage

  ## S4 method for signature 'kcca'
pairs(x, which=NULL, project=NULL, oma=NULL, ...)

Arguments

x
an object of class "kcca"

which
Index numbers of dimensions of (projected) input space to plot, default is to plot all dimensions.

project
Projection object for which a predict method exists, e.g., the result of prcomp.

oma
Outer margin.

...
Passed to the plot method.

Details

This works only for "kcca" objects, no method is available for "kccasimple" objects.

Author(s)

Friedrich Leisch

[Package flexclust version 1.3-5 Index]