plot-methods {flexclust}	R Documentation

Methods for Function plot in Package ‘flexclust’

Description

Plot the neighbourhood graph of a cluster solution together with projected data points.

Usage

  ## S4 method for signature 'kcca,missing'
plot(x, y, which=1:2, project=NULL,
         data=NULL, points=TRUE, hull=TRUE, hull.args=NULL, 
         number = TRUE, simlines=TRUE,
         lwd=1, maxlwd=8*lwd, cex=1.5, numcol=FALSE, nodes=16,
         add=FALSE, xlab="", ylab="", xlim = NULL,
         ylim = NULL, pch=NULL, col=NULL, ...)

Arguments

x
an object of class "kcca"

y
not used

which
Index numbers of dimensions of (projected) input space to plot.

project
Projection object for which a predict method exists, e.g., the result of prcomp.

data
Data to include in plot. If the cluster object x was created with save.data=TRUE, then these are used by default.

points
Logical, shall data points be plotted (if available)?

hull
If TRUE, then hulls of the data are plotted (if available). Can either be a logical value, one of the strings "convex" (the default) or "ellipse", or a function for plotting the hulls.

hull.args
A list of arguments for the hull function.

number
Logical, plot number labels in nodes of graph?

numcol, cex
Color and size of number labels in nodes of graph. If numcol is logical, it switches between black and the color of the clusters, else it is taken as a vector of colors.

nodes
Plotting symbol to use for nodes if no numbers are drawn.

simlines
Logical, plot edges of graph?

lwd, maxlwd
Numerical, thickness of lines.

add
Logical, add to existing plot?

xlab, ylab
Axis labels.

xlim, ylim
Axis range.

pch, col, ...
Plotting symbols and colors for data points.

Details

This works only for "kcca" objects, no method is available for "kccasimple" objects.

Author(s)

Friedrich Leisch

References

Friedrich Leisch. Visualizing cluster analysis and finite mixture models. In Chun houh Chen, Wolfgang Haerdle, and Antony Unwin, editors, Handbook of Data Visualization, Springer Handbooks of Computational Statistics. Springer Verlag, 2008.

[Package flexclust version 1.3-5 Index]