predict-methods {flexclust}	R Documentation

Predict Cluster Membership

Description

Return either the cluster membership of training data or predict for new data.

Usage

## S4 method for signature 'kccasimple'
predict(object, newdata, ...)
## S4 method for signature 'flexclust,ANY'
clusters(object, newdata, ...)

Arguments

object
Object of class inheriting from "flexclust".

newdata
An optional data matrix with the same number of columns as the cluster centers. If omitted, the fitted values are used.

...
Currently not used.

Details

clusters can be used on any object of class "flexclust" and returns the cluster memberships of the training data.

predict can be used only on objects of class "kcca" (which inherit from "flexclust"). If no newdata argument is specified, the function is identical to clusters, if newdata is specified, then cluster memberships for the new data are predicted. clusters(object, newdata, ...) is an alias for predict(object, newdata, ...).

Author(s)

Friedrich Leisch

[Package flexclust version 1.3-5 Index]