priceFeature {flexclust}	R Documentation

Artificial 2d Market Segment Data

Description

Simple artificial 2-dimensional data to demonstrate clustering for market segmentation. One dimension is the hypothetical feature sophistication (or performance or quality, etc) of a product, the second dimension the price customers are willing to pay for the product.

Usage

priceFeature(n, which=c("2clust", "3clust", "5clust",
                        "ellipse", "triangle", "circle", "square",
                        "largesmall"))

Arguments

n
sample size

which
shape of data set

References

Sara Dolnicar and Friedrich Leisch. Evaluation of structure and reproducibility of cluster solutions using the bootstrap. Marketing Letters, 21:83-101, 2010.

Examples

plot(priceFeature(200, "2clust"))
plot(priceFeature(200, "3clust"))
plot(priceFeature(200, "5clust"))
plot(priceFeature(200, "ell"))
plot(priceFeature(200, "tri"))
plot(priceFeature(200, "circ"))
plot(priceFeature(200, "square"))
plot(priceFeature(200, "largesmall"))

[Package flexclust version 1.3-5 Index]