volunteers {flexclust}	R Documentation

Motivation of Australian Volunteers

Description

Part of an Australian survey on motivation of volunteers to work for non-profit organisations like Red Cross, State Emergency Service, Rural Fire Service, Surf Life Saving, Rotary, Parents and Citizens Associations, etc..

Usage

data(volunteers)

Format

A data frame with 1415 observations on the following 21 variables: age and gender of respondents plus 19 binary motivation items (1 applies/ 0 does not apply).

GENDER

Gender of respondent.

AGEG

Age group, a factor with categorized age of respondents.

meet.people

I can meet different types of people.

no.one.else

There is no-one else to do the work.

example

It sets a good example for others.

socialise

I can socialise with people who are like me.

help.others

It gives me the chance to help others.

give.back

I can give something back to society.

career

It will help my career prospects.

lonely

It makes me feel less lonely.

active

It keeps me active.

community

It will improve my community.

cause

I can support an important cause.

faith

I can put faith into action.

services

I want to maintain services that I may use one day.

children

My children are involved with the organisation.

good.job

I feel like I am doing a good job.

benefited

I know someone who has benefited from the organisation.

network

I can build a network of contacts.

recognition

I can gain recognition within the community.

mind.off

It takes my mind off other things.

Source

Institute for Innovation in Business and Social Research, University of Wollongong, NSW, Australia

Melanie Randle, Friedrich Leisch, and Sara Dolnicar. Competition or collaboration? The effect of non-profit brand image on volunteer recruitment strategy. Journal of Brand Management, 2013. Accepted for publication.

[Package flexclust version 1.3-5 Index]